+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* [#8155](https://github.com/rubocop-hq/rubocop/pull/8155): Add helper constants containing nonmutating method names for several core classes. ([@fatkodima][])
 * [#8111](https://github.com/rubocop-hq/rubocop/pull/8111): Add auto-correct for `Style/StructInheritance`. ([@tejasbubane][])
 * [#8113](https://github.com/rubocop-hq/rubocop/pull/8113): Let `expect_offense` templates add variable-length whitespace with `_{foo}`. ([@eugeneius][])
 * [#8148](https://github.com/rubocop-hq/rubocop/pull/8148): Support autocorrection for `Style/MultilineTernaryOperator`. ([@koic][])
@@ -4597,3 +4598,4 @@
 [@andrykonchin]: https://github.com/andrykonchin
 [@avrusanov]: https://github.com/avrusanov
 [@mauro-oto]: https://github.com/mauro-oto
+[@fatkodima]: https://github.com/fatkodima